refactor: APIClient 생성 및 API 로직 리팩토링
