+++ src/_apis/client/mainClient.ts
@@ -1,0 +1,6 @@
+import BASE_URL from '../baseUrl';
+import APIClient from './APIClient';
+
+const mainClient = new APIClient(BASE_URL);
+
+export default mainClient;
