+++ src/_mocks/handlers/subscription.ts
@@ -1,13 +1,15 @@
-import API_ROUTES from '@/_apis/routes';
+import API_ROUTES from '@/_apis/constants/routes';
 import { http, HttpResponse } from 'msw';
 
 export const subscriptionHandlers = [
   http.post(API_ROUTES.post_subscribe, () => {
+    console.log('handle subscribe');
     return HttpResponse.json(null, {
       status: 200,
     });
   }),
   http.post(API_ROUTES.verify_email, () => {
+    console.log('handle verify email');
     return HttpResponse.json(null, {
       status: 200,
     });
